refactor(docker): move services tests.
